Move `check_path` to utils
